requirements.txt
# Biblioteca para interagir com a API da OpenAI
openai
# Biblioteca para indexação e busca vetorial
faiss-cpu
# Biblioteca para operações numéricas
numpy
# Biblioteca para manipulação de arquivos PDF
PyPDF2
# Biblioteca para carregar variáveis de ambiente de um arquivo .env
python-dotenv